Fix facets search in MultiFacetedSearchForm & fix ProductIndex price type (DecimalField represented in index such as string instead float val)

+++ oscar/apps/search/search_indexes.py
@@ -14,7 +14,7 @@ class ProductIndex(indexes.SearchIndex, indexes.Indexable):
 
     # Fields for faceting
     category = indexes.CharField(null=True, faceted=True)
-    price = indexes.DecimalField(null=True, faceted=True)
+    price = indexes.FloatField(null=True, faceted=True)
     num_in_stock = indexes.IntegerField(null=True, faceted=True)
 
     date_created = indexes.DateTimeField(model_attr='date_created')
